requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
supabase
python-dotenv
groq
pydantic
functions-framework
